feat: adding helthcheck route
