+++ routes/healthcheck_routes.py
@@ -1,0 +1,34 @@
+from flask import Blueprint
+from flasgger import swag_from
+
+# Routes for Healthcheck
+class HealthcheckRoutes(Blueprint):
+    def __init__(self):
+        super().__init__('healthcheck', __name__)
+        self.register_routes()
+
+    # Routes
+    def register_routes(self):
+        self.route('/healthcheck', methods=['GET'])(self.healthcheck)
+    
+  
+    @swag_from({
+        'tags': ['Health'],
+        'responses': {
+            200: {
+                'description': 'Service is up and running',
+                'schema': {
+                    'type': 'object',
+                    'properties': {
+                        'status': {'type': 'string'}
+                    }
+                }
+            },
+            500: {
+                'description': 'Internal server error'
+            }
+        }
+    })
+    # Veriifying a healthcheck
+    def healthcheck(self):
+        return jsonify({ 'status': 'up' }), 200
